+++ src/internal/connector/graph/middleware_test.go
@@ -292,3 +292,47 @@ func (suite *MiddlewareUnitSuite) TestBindExtractLimiterConfig() {
 		})
 	}
 }
+
+func (suite *MiddlewareUnitSuite) TestLimiterConsumption() {
+	ctx, flush := tester.NewContext()
+	defer flush()
+
+	// an unpopulated ctx should produce the default consumption
+	assert.Equal(suite.T(), defaultLC, ctxLimiterConsumption(ctx, defaultLC))
+
+	table := []struct {
+		name   string
+		n      int
+		expect int
+	}{
+		{
+			name:   "matches default",
+			n:      defaultLC,
+			expect: defaultLC,
+		},
+		{
+			name:   "default+1",
+			n:      defaultLC + 1,
+			expect: defaultLC + 1,
+		},
+		{
+			name:   "zero",
+			n:      0,
+			expect: defaultLC,
+		},
+		{
+			name:   "negative",
+			n:      -1,
+			expect: defaultLC,
+		},
+	}
+	for _, test := range table {
+		suite.Run(test.name, func() {
+			t := suite.T()
+
+			tctx := ConsumeNTokens(ctx, test.n)
+			lc := ctxLimiterConsumption(tctx, defaultLC)
+			assert.Equal(t, test.expect, lc)
+		})
+	}
+}