+++ src/internal/connector/graph_connector_onedrive_test.go
@@ -16,6 +16,7 @@ import (
 
 	"github.com/alcionai/corso/src/internal/common/ptr"
 	"github.com/alcionai/corso/src/internal/connector/graph"
+	odConsts "github.com/alcionai/corso/src/internal/connector/onedrive/consts"
 	"github.com/alcionai/corso/src/internal/connector/onedrive/metadata"
 	"github.com/alcionai/corso/src/internal/tester"
 	"github.com/alcionai/corso/src/internal/version"
@@ -109,7 +110,6 @@ var (
 	folderAName       = "folder-a"
 	folderBName       = "b"
 	folderNamedFolder = "folder"
-	rootFolder        = "root:"
 
 	fileAData = []byte(strings.Repeat("a", 33))
 	fileBData = []byte(strings.Repeat("b", 65))
@@ -255,7 +255,7 @@ func (c *onedriveCollection) withPermissions(perm permData) *onedriveCollection 
 		metaName = ""
 	}
 
-	if name == rootFolder {
+	if name == odConsts.RootPathDir {
 		return c
 	}
 
@@ -286,7 +286,7 @@ type itemData struct {
 	perms permData
 }
 
-type onedriveColInfo struct {
+type driveColInfo struct {
 	pathElements []string
 	perms        permData
 	files        []itemData
@@ -296,7 +296,7 @@ type onedriveColInfo struct {
 func testDataForInfo(
 	t *testing.T,
 	service path.ServiceType,
-	cols []onedriveColInfo,
+	cols []driveColInfo,
 	backupVersion int,
 ) []colInfo {
 	var res []colInfo
@@ -431,11 +431,6 @@ func (si suiteInfoImpl) Resource() Resource {
 // SharePoint shares most of its libraries implementation with OneDrive so we
 // only test simple things here and leave the more extensive testing to
 // OneDrive.
-//
-// TODO(ashmrtn): SharePoint doesn't have permissions backup/restore enabled
-// right now. Adjust the tests here when that is enabled so we have at least
-// basic assurances that it's doing the right thing. We can leave the more
-// extensive permissions tests to OneDrive as well.
 
 type GraphConnectorSharePointIntegrationSuite struct {
 	tester.Suite
@@ -484,6 +479,23 @@ func (suite *GraphConnectorSharePointIntegrationSuite) SetupSuite() {
 
 func (suite *GraphConnectorSharePointIntegrationSuite) TestRestoreAndBackup_MultipleFilesAndFolders_NoPermissions() {
 	testRestoreAndBackupMultipleFilesAndFoldersNoPermissions(suite, version.Backup)
+}
+
+func (suite *GraphConnectorSharePointIntegrationSuite) TestPermissionsRestoreAndBackup() {
+	testPermissionsRestoreAndBackup(suite, version.Backup)
+}
+
+func (suite *GraphConnectorSharePointIntegrationSuite) TestPermissionsBackupAndNoRestore() {
+	testPermissionsBackupAndNoRestore(suite, version.Backup)
+}
+
+func (suite *GraphConnectorSharePointIntegrationSuite) TestPermissionsInheritanceRestoreAndBackup() {
+	testPermissionsInheritanceRestoreAndBackup(suite, version.Backup)
+}
+
+func (suite *GraphConnectorSharePointIntegrationSuite) TestRestoreFolderNamedFolderRegression() {
+	// No reason why it couldn't work with previous versions, but this is when it got introduced.
+	testRestoreFolderNamedFolderRegression(suite, version.All8MigrateUserPNToID)
 }
 
 // ---------------------------------------------------------------------------
@@ -631,39 +643,39 @@ func testRestoreAndBackupMultipleFilesAndFoldersNoPermissions(
 		suite.BackupResourceOwner())
 
 	rootPath := []string{
-		"drives",
+		odConsts.DrivesPathDir,
 		driveID,
-		rootFolder,
+		odConsts.RootPathDir,
 	}
 	folderAPath := []string{
-		"drives",
+		odConsts.DrivesPathDir,
 		driveID,
-		rootFolder,
+		odConsts.RootPathDir,
 		folderAName,
 	}
 	subfolderBPath := []string{
-		"drives",
+		odConsts.DrivesPathDir,
 		driveID,
-		rootFolder,
+		odConsts.RootPathDir,
 		folderAName,
 		folderBName,
 	}
 	subfolderAPath := []string{
-		"drives",
+		odConsts.DrivesPathDir,
 		driveID,
-		rootFolder,
+		odConsts.RootPathDir,
 		folderAName,
 		folderBName,
 		folderAName,
 	}
 	folderBPath := []string{
-		"drives",
+		odConsts.DrivesPathDir,
 		driveID,
-		rootFolder,
+		odConsts.RootPathDir,
 		folderBName,
 	}
 
-	cols := []onedriveColInfo{
+	cols := []driveColInfo{
 		{
 			pathElements: rootPath,
 			files: []itemData{
@@ -776,38 +788,38 @@ func testPermissionsRestoreAndBackup(suite oneDriveSuite, startVersion int) {
 	folderCName := "folder-c"
 
 	rootPath := []string{
-		"drives",
+		odConsts.DrivesPathDir,
 		driveID,
-		rootFolder,
+		odConsts.RootPathDir,
 	}
 	folderAPath := []string{
-		"drives",
+		odConsts.DrivesPathDir,
 		driveID,
-		rootFolder,
+		odConsts.RootPathDir,
 		folderAName,
 	}
 	folderBPath := []string{
-		"drives",
+		odConsts.DrivesPathDir,
 		driveID,
-		rootFolder,
+		odConsts.RootPathDir,
 		folderBName,
 	}
 	// For skipped test
 	// subfolderAPath := []string{
-	// 	"drives",
+	// 	odConsts.DrivesPathDir,
 	// 	driveID,
-	// 	rootFolder,
+	// 	odConsts.RootPathDir,
 	// 	folderBName,
 	// 	folderAName,
 	// }
 	folderCPath := []string{
-		"drives",
+		odConsts.DrivesPathDir,
 		driveID,
-		rootFolder,
+		odConsts.RootPathDir,
 		folderCName,
 	}
 
-	cols := []onedriveColInfo{
+	cols := []driveColInfo{
 		{
 			pathElements: rootPath,
 			files: []itemData{
@@ -939,9 +951,10 @@ func testPermissionsRestoreAndBackup(suite oneDriveSuite, startVersion int) {
 	}
 
 	expected := testDataForInfo(suite.T(), suite.BackupService(), cols, version.Backup)
+	bss := suite.BackupService().String()
 
 	for vn := startVersion; vn <= version.Backup; vn++ {
-		suite.Run(fmt.Sprintf("Version%d", vn), func() {
+		suite.Run(fmt.Sprintf("%s-Version%d", bss, vn), func() {
 			t := suite.T()
 			// Ideally this can always be true or false and still
 			// work, but limiting older versions to use emails so as
@@ -984,12 +997,12 @@ func testPermissionsBackupAndNoRestore(suite oneDriveSuite, startVersion int) {
 		suite.Service(),
 		suite.BackupResourceOwner())
 
-	inputCols := []onedriveColInfo{
+	inputCols := []driveColInfo{
 		{
 			pathElements: []string{
-				"drives",
+				odConsts.DrivesPathDir,
 				driveID,
-				rootFolder,
+				odConsts.RootPathDir,
 			},
 			files: []itemData{
 				{
@@ -1005,12 +1018,12 @@ func testPermissionsBackupAndNoRestore(suite oneDriveSuite, startVersion int) {
 		},
 	}
 
-	expectedCols := []onedriveColInfo{
+	expectedCols := []driveColInfo{
 		{
 			pathElements: []string{
-				"drives",
+				odConsts.DrivesPathDir,
 				driveID,
-				rootFolder,
+				odConsts.RootPathDir,
 			},
 			files: []itemData{
 				{
@@ -1023,9 +1036,10 @@ func testPermissionsBackupAndNoRestore(suite oneDriveSuite, startVersion int) {
 	}
 
 	expected := testDataForInfo(suite.T(), suite.BackupService(), expectedCols, version.Backup)
+	bss := suite.BackupService().String()
 
 	for vn := startVersion; vn <= version.Backup; vn++ {
-		suite.Run(fmt.Sprintf("Version%d", vn), func() {
+		suite.Run(fmt.Sprintf("%s-Version%d", bss, vn), func() {
 			t := suite.T()
 			input := testDataForInfo(t, suite.BackupService(), inputCols, vn)
 
@@ -1073,34 +1087,34 @@ func testPermissionsInheritanceRestoreAndBackup(suite oneDriveSuite, startVersio
 	folderCName := "empty"
 
 	rootPath := []string{
-		"drives",
+		odConsts.DrivesPathDir,
 		driveID,
-		rootFolder,
+		odConsts.RootPathDir,
 	}
 	folderAPath := []string{
-		"drives",
+		odConsts.DrivesPathDir,
 		driveID,
-		rootFolder,
+		odConsts.RootPathDir,
 		folderAName,
 	}
 	subfolderAAPath := []string{
-		"drives",
+		odConsts.DrivesPathDir,
 		driveID,
-		rootFolder,
+		odConsts.RootPathDir,
 		folderAName,
 		folderAName,
 	}
 	subfolderABPath := []string{
-		"drives",
+		odConsts.DrivesPathDir,
 		driveID,
-		rootFolder,
+		odConsts.RootPathDir,
 		folderAName,
 		folderBName,
 	}
 	subfolderACPath := []string{
-		"drives",
+		odConsts.DrivesPathDir,
 		driveID,
-		rootFolder,
+		odConsts.RootPathDir,
 		folderAName,
 		folderCName,
 	}
@@ -1150,7 +1164,7 @@ func testPermissionsInheritanceRestoreAndBackup(suite oneDriveSuite, startVersio
 	// 	   - inherted-permission-file
 	//     - empty-permission-file (empty/empty might have interesting behavior)
 
-	cols := []onedriveColInfo{
+	cols := []driveColInfo{
 		{
 			pathElements: rootPath,
 			files:        []itemData{},
@@ -1199,9 +1213,10 @@ func testPermissionsInheritanceRestoreAndBackup(suite oneDriveSuite, startVersio
 	}
 
 	expected := testDataForInfo(suite.T(), suite.BackupService(), cols, version.Backup)
+	bss := suite.BackupService().String()
 
 	for vn := startVersion; vn <= version.Backup; vn++ {
-		suite.Run(fmt.Sprintf("Version%d", vn), func() {
+		suite.Run(fmt.Sprintf("%s-Version%d", bss, vn), func() {
 			t := suite.T()
 			// Ideally this can always be true or false and still
 			// work, but limiting older versions to use emails so as
@@ -1246,25 +1261,25 @@ func testRestoreFolderNamedFolderRegression(
 		suite.BackupResourceOwner())
 
 	rootPath := []string{
-		"drives",
+		odConsts.DrivesPathDir,
 		driveID,
-		rootFolder,
+		odConsts.RootPathDir,
 	}
 	folderFolderPath := []string{
-		"drives",
+		odConsts.DrivesPathDir,
 		driveID,
-		rootFolder,
+		odConsts.RootPathDir,
 		folderNamedFolder,
 	}
 	subfolderPath := []string{
-		"drives",
+		odConsts.DrivesPathDir,
 		driveID,
-		rootFolder,
+		odConsts.RootPathDir,
 		folderNamedFolder,
 		folderBName,
 	}
 
-	cols := []onedriveColInfo{
+	cols := []driveColInfo{
 		{
 			pathElements: rootPath,
 			files: []itemData{
@@ -1313,9 +1328,10 @@ func testRestoreFolderNamedFolderRegression(
 	}
 
 	expected := testDataForInfo(suite.T(), suite.BackupService(), cols, version.Backup)
+	bss := suite.BackupService().String()
 
 	for vn := startVersion; vn <= version.Backup; vn++ {
-		suite.Run(fmt.Sprintf("Version%d", vn), func() {
+		suite.Run(fmt.Sprintf("%s-Version%d", bss, vn), func() {
 			t := suite.T()
 			input := testDataForInfo(t, suite.BackupService(), cols, vn)