+++ src/internal/connector/onedrive/consts/consts.go
@@ -1,0 +1,10 @@
+package onedrive
+
+const (
+	// const used as the root dir for the drive portion of a path prefix.
+	// eg: tid/onedrive/ro/files/drives/driveid/...
+	DrivesPathDir = "drives"
+	// const used as the root-of-drive dir for the drive portion of a path prefix.
+	// eg: tid/onedrive/ro/files/drives/driveid/root:/...
+	RootPathDir = "root:"
+)