+++ src/pkg/control/options.go
@@ -101,6 +101,14 @@ type Toggles struct {
 	// DisableIncrementals prevents backups from using incremental lookups,
 	// forcing a new, complete backup of all data regardless of prior state.
 	DisableIncrementals bool `json:"exchangeIncrementals,omitempty"`
+	// DisableDelta prevents backups from using delta based lookups,
+	// forcing a backup by enumerating all items. This is different
+	// from DisableIncrementals in that this does not even makes use of
+	// delta endpoints with or without a delta token. This is necessary
+	// when the user has filled up the mailbox storage available to the
+	// user as Microsoft prevents the API from being able to make calls
+	// to delta endpoints.
+	DisableDelta bool `json:"exchangeDelta,omitempty"`
 	// ExchangeImmutableIDs denotes whether Corso should store items with
 	// immutable Exchange IDs. This is only safe to set if the previous backup for
 	// incremental backups used immutable IDs or if a full backup is being done.